FixedVec is extended by try_grow
try_grow to increase the capacity of the pinned vector with default additional amount defined by the specific implementation. It is not guaranteed that the pinned vector will succeed while keeping the elements pinned to their locations; and hence, will return a result.

Currently FixedVec cannot grow and returns the error.